requirements.txt
APScheduler==3.5.1
certifi==2018.1.18
chardet==3.0.4
idna==2.6
oauthlib==2.0.7
PySocks==1.6.8
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.6.0
tzlocal==1.5.1
urllib3==1.22
